feat: add browser and window management APIs
- Introduced new `browser` and `window` modules to manage browser plugins and window auto-start features.
- Added routes for browser plugin installation, status checking, and updates.
- Implemented window auto-start check, enable, and disable functionalities.
- Refactored existing code to streamline browser-related operations and improve organization.
